Refactor the dashboard page
